Se realiza modificación en appContext, para integrar el context y el reducer de profile
