+++ src/main.jsx
@@ -2,11 +2,10 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import AppRouter from './router/AppRouter'
 import AppContextProvider from './context/AppContext';
-import { ProfileProvider } from './context/ProfileContext'
 import '../src/index.css'
 
 createRoot(document.getElementById('root')).render(
-  <StrictMode>
+  <AppContextProvider initialProfileData={{ /* datos iniciales del perfil */ }}>
     <AppRouter />
-  </StrictMode>,
+  </AppContextProvider>,
 )
